Removed unneeded package .imagecrop.
