+++ android-image-zoom-crop/src/main/java/io/togoto/imagezoomcrop/cropoverlay/edge/Edge.java
@@ -11,7 +11,7 @@
  * governing permissions and limitations under the License. 
 */
 
-package io.togoto.imagezoomcrop.imagecrop.cropoverlay.edge;
+package io.togoto.imagezoomcrop.cropoverlay.edge;
 
 /**
  * Enum representing an edge in the crop window.
