+++ android-image-zoom-crop/src/main/java/io/togoto/imagezoomcrop/cropoverlay/utils/PaintUtil.java
@@ -11,7 +11,7 @@
  * governing permissions and limitations under the License. 
  */
 
-package io.togoto.imagezoomcrop.imagecrop.cropoverlay.utils;
+package io.togoto.imagezoomcrop.cropoverlay.utils;
 
 import android.content.Context;
 import android.graphics.Color;
