+++ android-image-zoom-crop/src/main/java/io/togoto/imagezoomcrop/photoview/DefaultOnDoubleTapListener.java
@@ -1,4 +1,4 @@
-package io.togoto.imagezoomcrop.imagecrop.photoview;
+package io.togoto.imagezoomcrop.photoview;
 
 import android.graphics.RectF;
 import android.view.GestureDetector;
