+++ android-image-zoom-crop/src/main/java/io/togoto/imagezoomcrop/photoview/gestures/CupcakeGestureDetector.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package io.togoto.imagezoomcrop.imagecrop.photoview.gestures;
+package io.togoto.imagezoomcrop.photoview.gestures;
 
 import android.content.Context;
 import android.util.Log;
